Fix xsdgen to support 3gpp compliant XML files
`xsdgen` was failing for the `xsd` present in:
https://www.3gpp.org/ftp/specs/archive/32_series/32.435#measCollec

As there was no entry for the `xsd.AnySimpleType`, it was failing
to generate the structs, marshalling, unmarshalling methods, etc.
for the same.

With reference to:
```
<simpleType name="measResultType">
        <union memberTypes="float">
            <simpleType>
                <restriction base="string">
                    <enumeration value="NIL" />
                </restriction>
            </simpleType>
        </union>
</simpleType>
```
The type could be `float` as well as string "NIL". So, current
implementation is as follows:

If the XML contains NIL or "NIL", it'll considered as 0.0 (`float64`)
during unmarshalling. This is a custom implementation and shouldn't
be used for other XML files which don't respect the `xsd` file taken
into consideration here.